move/rename errors.py
