+++ armory/art_experimental/attacks/carla_obj_det_patch.py
@@ -5,7 +5,7 @@ from art.attacks.evasion import RobustDPatch
 import cv2
 import numpy as np
 
-from armory.errors import ExternalRepoImport
+from armory.utils.external_repo import ExternalRepoImport
 
 with ExternalRepoImport(
     repo="colour-science/colour@v0.3.16",
