+++ armory/utils/external_repo.py
@@ -1,6 +1,7 @@
 """
 Utils to pull external repos for evaluation
 """
+import contextlib
 import os
 import tarfile
 import shutil
@@ -12,6 +13,27 @@ import requests
 from armory import paths
 from armory.configuration import get_verify_ssl
 from armory.logs import log
+
+
+class ExternalRepoImport(contextlib.AbstractContextManager):
+    def __init__(self, repo="", experiment=""):
+        super().__init__()
+        url = f"https://github.com/{repo}"
+        name = repo.split("/")[-1].split("@")[0]
+        self.error_message = "\n".join(
+            [
+                f"{name} is an external repo.",
+                f"Please download from {url} and place on local environment PYTHONPATH",
+                "    OR place in experimental config `external_github_repo` field.",
+                f"    See scenario_configs/{experiment} for an example.",
+            ]
+        )
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        if exc_type is not None and issubclass(exc_type, ImportError):
+            log.error(self.error_message)
+            return False
+        return True
 
 
 def add_path(path, include_parent=False, index=1):
